First attempt at generation using brute force, unsuccessful.
I wasn't expecting this to work, but I wanted to try something quickly. The Generate method
attempts to loop over each spot, and pick a random number. After 20 attempts at picking a number (to account for randomness) it resets the board and tries again. I was hoping it would still be possible to generate a board in 20-30 seconds by luck, but unfortunatly not.

The next attempt will be using a back-tracking algorithm to check all possibilites. Which should be considerably faster.