Backtracking version of the Generator appears to be working.
But the Validator does not. An invalid test case that was generated shows the Validator
is accepting invalid boards... I'm guessing there is something in wrong with the validateRow
method... if it was a problem in the build methods, at least one of the conditions should have
failed and caused the entire thing to go. Interesting.